requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI and ML
voyageai>=0.3.4
sentence-transformers==2.2.2
tiktoken==0.5.1

# Vector database
pinecone-client==2.2.4
faiss-cpu==1.7.4

# Document processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0

# Data processing
numpy==1.24.3
pandas==2.1.3

# Utilities
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1